switch to Erlang like actor model

+++ src/crawler/net_asset_value.rs
@@ -1,4 +1,5 @@
 use chrono::{Duration, NaiveDate};
+use rust_decimal::Decimal;
 use serde::Deserialize;
 use soup::prelude::*;
 
@@ -15,7 +16,7 @@ async fn get_page(
     start: &str,
     end: &str,
     page: u32,
-) -> Result<(Vec<(NaiveDate, f64)>, bool), Box<dyn std::error::Error>> {
+) -> Result<(Vec<(NaiveDate, Decimal)>, bool), Box<dyn std::error::Error>> {
     let body = reqwest::get(&format!(
         "http://fund.eastmoney.com/f10/F10DataApi.aspx?\
          type=lsjz&code={}&page={}&per=20&sdate={}&edate={}",
@@ -47,7 +48,7 @@ pub async fn get_history(
     code: &str,
     start: &str,
     end: &str,
-) -> Result<Vec<(NaiveDate, f64)>, Box<dyn std::error::Error>> {
+) -> Result<Vec<(NaiveDate, Decimal)>, Box<dyn std::error::Error>> {
     let end = &(end.parse::<NaiveDate>()? - Duration::days(1)).to_string();
     let mut history = vec![];
     let mut page = 1;
